move casts to crypto_util
