+++ tests/test_file_metadata.py
@@ -12,7 +12,7 @@ def test_file_hashes():
         f.write(b'\x11' * 1001)
         f.seek(0)
 
-        h_out = file_hashes(f)
+        h_out = file_hashes(f, 2**23)
         expected_data = b'\x11' * 1001
         h_expected = hashlib.sha256(expected_data).digest()
